requirements.txt
torch==1.8.1+cu111
torchvision==0.9.1+cu111
timm==0.4.5
tqdm
scikit-image==0.18.1
scikit-learn==1.0.1
pytorch-msssim==0.2.1
Pillow==7.0.0
opencv-python==4.5.5.64
jupyter
compressai==1.2.0
pandas
